+++ docs/changes.rst
@@ -8,6 +8,7 @@ Released on ? 2019.
 
 * Added separate Celery queue for notifications.
 * Use consistent look with application for API browsing.
+* Include approved stats in the reports.
 
 weblate 3.6.1
 -------------